feat(ui): add MCP install PR flow from repositories page
Extract the GitHub MCP config PR wizard into McpConfigPrWizard for reuse
from onboarding and repo management. Add a wide modal and menu action
under the repositories triple-dot menu so users can preview diffs and
raise PRs for already-ingested GitHub repos (pre-selected by clone URL).

Refs CTX-32.
