+++ apps/ui/src/components/ui/Modal.tsx
@@ -1,7 +1,12 @@
 "use client"
-import { composeRenderProps, ModalOverlay, Modal as RACModal } from "react-aria-components"
 import type { ModalOverlayProps } from "react-aria-components"
+import {
+  composeRenderProps,
+  ModalOverlay,
+  Modal as RACModal,
+} from "react-aria-components"
 import { tv } from "tailwind-variants"
+import { cn } from "@/lib/utils"
 
 const overlayStyles = tv({
   base: "absolute top-0 left-0 w-full h-(--page-height) isolate z-20 bg-black/[45%] text-center backdrop-blur-sm",
@@ -27,14 +32,16 @@ const modalStyles = tv({
   },
 })
 
-export function Modal(props: ModalOverlayProps) {
-  const { children, ...overlayProps } = props
+type ModalSize = "default" | "wide"
+
+export function Modal(props: ModalOverlayProps & { size?: ModalSize }) {
+  const { children, size = "default", ...overlayProps } = props
   return (
     <ModalOverlay {...overlayProps} className={overlayStyles}>
       <div className="sticky top-0 left-0 w-full h-(--visual-viewport-height) flex items-center justify-center box-border">
         <RACModal
           className={composeRenderProps(undefined, (_, r) =>
-            modalStyles(r),
+            cn(modalStyles(r), size === "wide" && "max-w-[min(94vw,960px)]"),
           )}
         >
           {children}
